Use Int64 connectivity in .vtu output for meshes above the Int32 limit
The PARAVIEW (.vtu) writer stored connectivity and offsets as Int32, and
the cumulative connectivity counter of CParallelDataSorter was an int.
Meshes with more than 2^31-1 connectivity entries (e.g. more than ~268M
hexahedra or ~358M prisms) got overflowed offsets and unreadable files.

- Write connectivity and offsets as Int64 when the global connectivity
  size does not fit in Int32; smaller meshes keep Int32 (files unchanged).
- Store nElemConn_Send/nElemConn_Cum as unsigned long.
- Write NumberOfPoints/NumberOfCells without casting to int.

+++ SU2_CFD/src/output/filewriter/CParallelDataSorter.cpp
@@ -59,8 +59,8 @@ CParallelDataSorter::CParallelDataSorter(CConfig *config, const vector<string> &
   nPoint_Recv = new int[size+1]();
   nElem_Send  = new int[size+1]();
   nElem_Cum  = new int[size+1]();
-  nElemConn_Send = new int[size+1]();
-  nElemConn_Cum = new int[size+1]();
+  nElemConn_Send = new unsigned long[size+1]();
+  nElemConn_Cum = new unsigned long[size+1]();
 
   nElemPerType.fill(0);
   nElemPerTypeGlobal.fill(0);
@@ -397,7 +397,7 @@ void CParallelDataSorter::SetTotalElements(){
   nElem_Cum[0] = 0; nElemConn_Cum[0] = 0;
   for (int ii=1; ii <= size; ii++) {
     nElem_Send[ii]     = int(nElem);
-    nElemConn_Send[ii] = int(nConn);
+    nElemConn_Send[ii] = nConn;
     nElem_Cum[ii] = 0;
     nElemConn_Cum[ii] = 0;
   }
@@ -407,8 +407,8 @@ void CParallelDataSorter::SetTotalElements(){
   SU2_MPI::Alltoall(&(nElem_Send[1]), 1, MPI_INT,
                     &(nElem_Cum[1]), 1, MPI_INT, SU2_MPI::GetComm());
 
-  SU2_MPI::Alltoall(&(nElemConn_Send[1]), 1, MPI_INT,
-                    &(nElemConn_Cum[1]), 1, MPI_INT, SU2_MPI::GetComm());
+  SU2_MPI::Alltoall(&(nElemConn_Send[1]), 1, MPI_UNSIGNED_LONG,
+                    &(nElemConn_Cum[1]), 1, MPI_UNSIGNED_LONG, SU2_MPI::GetComm());
 
   /*--- Put the counters into cumulative storage format. ---*/
